remove legacy namespace
