+++ packages/teams-js/src/private/index.ts
@@ -20,7 +20,6 @@ export {
 } from './privateAPIs';
 export { conversations } from './conversations';
 export { files } from './files';
-export { legacy } from './legacy';
 export { meetingRoom } from './meetingRoom';
 export { notifications } from './notifications';
 export { remoteCamera } from './remoteCamera';
