+++ packages/teams-js/test/private/teams.spec.ts
@@ -1,4 +1,4 @@
-import { teams } from '../../src/private';
+import { TeamInstanceParameters, teams } from '../../src/private';
 import { app } from '../../src/public';
 import { errorNotSupportedOnPlatform, FrameContexts } from '../../src/public/constants';
 import { _minRuntimeConfigToUninitialize } from '../../src/public/runtime';
@@ -159,6 +159,187 @@ describe('Testing teams capabillity', () => {
       it('teams.isSupported should return true if the runtime says teams is supported', () => {
         utils.setRuntimeConfig({ apiVersion: 1, supports: { teams: {} } });
         expect(teams.isSupported()).toBeTruthy();
+      });
+    });
+
+    describe('Testing teams.getUserJoinedTeams', () => {
+      it('should not allow calls before initialization', () => {
+        return expect(teams.fullTrust.joinedTeams.getUserJoinedTeams()).rejects.toThrowError(
+          'The library has not yet been initialized',
+        );
+      });
+      it('should throw error when fullTrust.joinedTeams is not supported', async () => {
+        await utils.initializeWithContext('content');
+        utils.setRuntimeConfig({ apiVersion: 1, supports: { teams: { fullTrust: {} } } });
+        expect(() => teams.fullTrust.joinedTeams.getUserJoinedTeams()).rejects.toEqual(errorNotSupportedOnPlatform);
+      });
+
+      it('should throw error when fullTrust is not supported', async () => {
+        await utils.initializeWithContext('content');
+        utils.setRuntimeConfig({ apiVersion: 1, supports: { teams: {} } });
+        expect(() => teams.fullTrust.joinedTeams.getUserJoinedTeams()).rejects.toEqual(errorNotSupportedOnPlatform);
+      });
+
+      it('should throw error when teams is not supported', async () => {
+        await utils.initializeWithContext('content');
+        utils.setRuntimeConfig({ apiVersion: 1, supports: {} });
+        expect(() => teams.fullTrust.joinedTeams.getUserJoinedTeams()).rejects.toEqual(errorNotSupportedOnPlatform);
+      });
+
+      it('should allow a valid optional parameter set to true', async () => {
+        await utils.initializeWithContext('content');
+        utils.setRuntimeConfig({ apiVersion: 1, supports: { teams: { fullTrust: { joinedTeams: {} } } } });
+
+        const promise = teams.fullTrust.joinedTeams.getUserJoinedTeams({
+          favoriteTeamsOnly: true,
+        } as TeamInstanceParameters);
+
+        const getUserJoinedTeamsMessage = utils.findMessageByFunc('getUserJoinedTeams');
+        expect(getUserJoinedTeamsMessage).not.toBeNull();
+        utils.respondToMessage(getUserJoinedTeamsMessage, {});
+        return expect(promise).resolves;
+      });
+
+      it('should allow a valid optional parameter set to false', async () => {
+        await utils.initializeWithContext('content');
+        utils.setRuntimeConfig({ apiVersion: 1, supports: { teams: { fullTrust: { joinedTeams: {} } } } });
+
+        const promise = teams.fullTrust.joinedTeams.getUserJoinedTeams({
+          favoriteTeamsOnly: false,
+        } as TeamInstanceParameters);
+
+        const getUserJoinedTeamsMessage = utils.findMessageByFunc('getUserJoinedTeams');
+        expect(getUserJoinedTeamsMessage).not.toBeNull();
+        utils.respondToMessage(getUserJoinedTeamsMessage, {});
+        return expect(promise).resolves;
+      });
+
+      it('should allow a missing optional parameter', async () => {
+        await utils.initializeWithContext('content');
+        utils.setRuntimeConfig({ apiVersion: 1, supports: { teams: { fullTrust: { joinedTeams: {} } } } });
+
+        const promise = teams.fullTrust.joinedTeams.getUserJoinedTeams();
+
+        const getUserJoinedTeamsMessage = utils.findMessageByFunc('getUserJoinedTeams');
+        expect(getUserJoinedTeamsMessage).not.toBeNull();
+        utils.respondToMessage(getUserJoinedTeamsMessage, {});
+        return expect(promise).resolves;
+      });
+
+      it('should allow a missing and valid optional parameter', async () => {
+        await utils.initializeWithContext('content');
+        utils.setRuntimeConfig({ apiVersion: 1, supports: { teams: { fullTrust: { joinedTeams: {} } } } });
+        const promise = teams.fullTrust.joinedTeams.getUserJoinedTeams({} as TeamInstanceParameters);
+
+        const getUserJoinedTeamsMessage = utils.findMessageByFunc('getUserJoinedTeams');
+        expect(getUserJoinedTeamsMessage).not.toBeNull();
+        utils.respondToMessage(getUserJoinedTeamsMessage, {});
+        return expect(promise).resolves;
+      });
+    });
+
+    describe('joinedTeams.isSupported function', () => {
+      const utils = new Utils();
+      it('joinedTeams.isSupported should return false if the runtime says joinedTeams is not supported', () => {
+        utils.setRuntimeConfig({ apiVersion: 1, supports: {} });
+        expect(teams.fullTrust.joinedTeams.isSupported()).not.toBeTruthy();
+      });
+
+      it('joinedTeams.isSupported should return false if the runtime says joinedTeams is not supported when teams is supported', () => {
+        utils.setRuntimeConfig({ apiVersion: 1, supports: { teams: {} } });
+        expect(teams.fullTrust.joinedTeams.isSupported()).not.toBeTruthy();
+      });
+
+      it('joinedTeams.isSupported should return false if the runtime says joinedTeams is not supported when teams and fullTrust is supported', () => {
+        utils.setRuntimeConfig({ apiVersion: 1, supports: { teams: { fullTrust: {} } } });
+        expect(teams.fullTrust.joinedTeams.isSupported()).not.toBeTruthy();
+      });
+
+      it('joinedTeams.isSupported should return true if the runtime says joinedTeams is supported', () => {
+        utils.setRuntimeConfig({ apiVersion: 1, supports: { teams: { fullTrust: { joinedTeams: {} } } } });
+        expect(teams.fullTrust.joinedTeams.isSupported()).toBeTruthy();
+      });
+    });
+
+    describe('teams.fullTrust.isSupported function', () => {
+      const utils = new Utils();
+      it('teams.fullTrust.isSupported should return false if the runtime says fullTrust is not supported', () => {
+        utils.setRuntimeConfig({ apiVersion: 1, supports: {} });
+        expect(teams.fullTrust.isSupported()).not.toBeTruthy();
+      });
+
+      it('teams.fullTrust.isSupported should return false if the runtime says fullTrust is not supported when teams is supported', () => {
+        utils.setRuntimeConfig({ apiVersion: 1, supports: { teams: {} } });
+        expect(teams.fullTrust.isSupported()).not.toBeTruthy();
+      });
+
+      it('teams.fullTrust.isSupported should return true if the runtime says fullTrust is supported', () => {
+        utils.setRuntimeConfig({ apiVersion: 1, supports: { teams: { fullTrust: {} } } });
+        expect(teams.fullTrust.isSupported()).toBeTruthy();
+      });
+    });
+
+    describe('getConfigSetting', () => {
+      it('should not allow calls before initialization', () => {
+        return expect(teams.fullTrust.getConfigSetting('key')).rejects.toThrowError(
+          'The library has not yet been initialized',
+        );
+      });
+
+      it('should throw error when fullTrust is not supported', async () => {
+        await utils.initializeWithContext('content');
+        utils.setRuntimeConfig({ apiVersion: 1, supports: { teams: {} } });
+        expect(() => teams.fullTrust.getConfigSetting('key')).rejects.toEqual(errorNotSupportedOnPlatform);
+      });
+
+      it('should throw error when teams is not supported', async () => {
+        await utils.initializeWithContext('content');
+        utils.setRuntimeConfig({ apiVersion: 1, supports: {} });
+        expect(() => teams.fullTrust.getConfigSetting('key')).rejects.toEqual(errorNotSupportedOnPlatform);
+      });
+
+      it('should allow a valid parameter', async () => {
+        await utils.initializeWithContext('content');
+
+        const promise = teams.fullTrust.getConfigSetting('key');
+
+        const getConfigSettingMessage = utils.findMessageByFunc('getConfigSetting');
+        expect(getConfigSettingMessage).not.toBeNull();
+        utils.respondToMessage(getConfigSettingMessage, {});
+        return expect(promise).resolves;
+      });
+
+      it('should allow an empty key', async () => {
+        await utils.initializeWithContext('content');
+
+        const promise = teams.fullTrust.getConfigSetting('');
+
+        const getConfigSettingMessage = utils.findMessageByFunc('getConfigSetting');
+        expect(getConfigSettingMessage).not.toBeNull();
+        utils.respondToMessage(getConfigSettingMessage, {});
+        expect(promise).resolves;
+      });
+
+      it('should allow a null key', async () => {
+        await utils.initializeWithContext('content');
+
+        const promise = teams.fullTrust.getConfigSetting(null);
+
+        const getConfigSettingMessage = utils.findMessageByFunc('getConfigSetting');
+        expect(getConfigSettingMessage).not.toBeNull();
+        utils.respondToMessage(getConfigSettingMessage, {});
+        expect(promise).resolves;
+      });
+
+      it('should allow an undefined key', async () => {
+        await utils.initializeWithContext('content');
+
+        const promise = teams.fullTrust.getConfigSetting(undefined);
+
+        const getConfigSettingMessage = utils.findMessageByFunc('getConfigSetting');
+        expect(getConfigSettingMessage).not.toBeNull();
+        utils.respondToMessage(getConfigSettingMessage, {});
+        expect(promise).resolves;
       });
     });
   });
